videos para explicar el flujo de llamadas a la api durante la ejecución del juego

+++ Assets/Scripts/APIScripts/ScoreResponse.cs
@@ -3,5 +3,5 @@
 public class ScoreResponse
 {
     public string name;
-    public int puntuacion;
+    public float puntuacion;
 }
